+++ src/components/Cocktail/CocktailIndex.vue
@@ -17,6 +17,23 @@
                             <input :id="'global-' + filter.id" v-model="activeFilters[filter.id]" type="checkbox" :value="filter.active" @change="updateRouterPath">
                             <label :for="'global-' + filter.id">{{ filter.name }}</label>
                         </div>
+                        <SaltRimDialog v-model="showSpecificIngredientsModal">
+                            <template #trigger>
+                                <a href="#" @click.prevent="showSpecificIngredientsModal = true">{{ $t('search.ingredients') }} ({{ activeFilters.specific_ingredients.length }})</a>
+                            </template>
+                            <template #dialog>
+                                <FilterIngredientsModal :title="$t('search.select-specific-ingredients')" :value="activeFilters.specific_ingredients" @close="updateSpecificIngredients"></FilterIngredientsModal>
+                            </template>
+                        </SaltRimDialog>
+                        <br>
+                        <SaltRimDialog v-model="showIgnoreIngredientsModal">
+                            <template #trigger>
+                                <a href="#" @click.prevent="showIgnoreIngredientsModal = true">{{ $t('search.ignore-ingredients') }} ({{ activeFilters.ignore_ingredients.length }})</a>
+                            </template>
+                            <template #dialog>
+                                <FilterIngredientsModal :title="$t('search.select-ingredients-to-ignore')" :value="activeFilters.ignore_ingredients" @close="updateIgnoredIngredients"></FilterIngredientsModal>
+                            </template>
+                        </SaltRimDialog>
                     </Refinement>
                     <Refinement v-if="refineCollections.length > 0" id="collection" v-model="activeFilters.collections" :title="$t('collections.title')" :refinements="refineCollections" @change="updateRouterPath"></Refinement>
                     <Refinement v-if="refineUserShelves.length > 0" id="user_shelves" v-model="activeFilters.user_shelves" :title="$t('public-shelves')" :refinements="refineUserShelves" @change="updateRouterPath"></Refinement>
@@ -111,6 +128,7 @@ import SaltRimDialog from './../Dialog/SaltRimDialog.vue'
 import qs from 'qs'
 import AppState from '../../AppState'
 import EmptyState from './../EmptyState.vue'
+import FilterIngredientsModal from '../Search/FilterIngredientsModal.vue'
 
 export default {
     components: {
@@ -122,7 +140,8 @@ export default {
         SaltRimDialog,
         CollectionDialog,
         Pagination,
-        EmptyState
+        EmptyState,
+        FilterIngredientsModal,
     },
     data() {
         return {
@@ -130,6 +149,8 @@ export default {
             showCreateNewCollectionDialog: false,
             isLoading: false,
             showRefinements: false,
+            showIgnoreIngredientsModal: false,
+            showSpecificIngredientsModal: false,
             cocktails: [],
             favorites: [],
             searchQuery: null,
@@ -185,7 +206,10 @@ export default {
                 total_ingredients: null,
                 missing_ingredients: null,
                 user_shelves: [],
-                users: []
+                users: [],
+                ignore_ingredients: [],
+                specific_ingredients: [],
+                ingredient_id: [],
             }
         }
     },
@@ -407,6 +431,9 @@ export default {
             this.activeFilters.is_public = state.filter && state.filter.is_public ? state.filter.is_public : null
             this.activeFilters.total_ingredients = state.filter && state.filter.total_ingredients ? state.filter.total_ingredients : null
             this.activeFilters.missing_ingredients = state.filter && state.filter.missing_ingredients ? state.filter.missing_ingredients : null
+            this.activeFilters.ignore_ingredients = state.filter && state.filter.ignore_ingredients ? String(state.filter.ignore_ingredients).split(',') : []
+            this.activeFilters.specific_ingredients = state.filter && state.filter.specific_ingredients ? String(state.filter.specific_ingredients).split(',') : []
+            this.activeFilters.ingredient_id = state.filter && state.filter.ingredient_id ? String(state.filter.ingredient_id).split(',') : []
             this.activeFilters.user_rating = state.filter && state.filter.user_rating_min ? state.filter.user_rating_min : null
             this.activeFilters.average_rating = state.filter && state.filter.average_rating_min ? state.filter.average_rating_min : null
             this.searchQuery = state.filter && state.filter.name ? state.filter.name : null
@@ -442,11 +469,13 @@ export default {
                 average_rating_min: this.activeFilters.average_rating ? this.activeFilters.average_rating : null,
                 total_ingredients: this.activeFilters.total_ingredients ? this.activeFilters.total_ingredients : null,
                 missing_ingredients: this.activeFilters.missing_ingredients ? this.activeFilters.missing_ingredients : null,
+                ignore_ingredients: this.activeFilters.ignore_ingredients.length > 0 ? this.activeFilters.ignore_ingredients.join(',') : null,
                 tag_id: this.activeFilters.tags.length > 0 ? this.activeFilters.tags.join(',') : null,
                 glass_id: this.activeFilters.glasses.length > 0 ? this.activeFilters.glasses.join(',') : null,
                 cocktail_method_id: this.activeFilters.methods.length > 0 ? this.activeFilters.methods.join(',') : null,
                 main_ingredient_id: this.activeFilters.main_ingredients.length > 0 ? this.activeFilters.main_ingredients.join(',') : null,
-                ingredient_id: this.activeFilters.ingredients.length > 0 ? this.activeFilters.ingredients.join(',') : null,
+                specific_ingredients: this.activeFilters.specific_ingredients.length > 0 ? this.activeFilters.specific_ingredients.join(',') : null,
+                ingredient_id: this.activeFilters.ingredient_id.length > 0 ? this.activeFilters.ingredient_id.join(',') : null,
                 collection_id: this.activeFilters.collections.length > 0 ? this.activeFilters.collections.join(',') : null,
                 user_shelves: this.activeFilters.user_shelves.length > 0 ? this.activeFilters.user_shelves.join(',') : null,
                 created_user_id: this.activeFilters.users.length > 0 ? this.activeFilters.users.join(',') : null,
@@ -458,6 +487,16 @@ export default {
             query.filter = Object.entries(filters).reduce((a,[k,v]) => (v === null || v === false ? a : (a[k]=v, a)), {})
 
             return query
+        },
+        updateIgnoredIngredients(e) {
+            this.activeFilters.ignore_ingredients = e.newFilters
+            this.showIgnoreIngredientsModal = false
+            this.updateRouterPath()
+        },
+        updateSpecificIngredients(e) {
+            this.activeFilters.specific_ingredients = e.newFilters
+            this.showSpecificIngredientsModal = false
+            this.updateRouterPath()
         },
         handleCollectionsDialogClosed() {
             this.showCreateNewCollectionDialog = false
@@ -493,7 +532,10 @@ export default {
                 abv: null,
                 total_ingredients: null,
                 user_shelves: [],
-                users: []
+                users: [],
+                ignore_ingredients: [],
+                specific_ingredients: [],
+                ingredient_id: [],
             }
 
             this.updateRouterPath()